Rename AddressField components
