+++ src/platform/user/profile/vet360/components/AddressField/AddressEditModal.jsx
@@ -1,10 +1,10 @@
 import React from 'react';
 
-import { FIELD_NAMES, ADDRESS_FORM_VALUES, USA } from '../../constants';
+import { FIELD_NAMES, ADDRESS_FORM_VALUES, USA } from 'vet360/constants';
 
 import Vet360EditModal from '../base/EditModal';
 
-import CopyMailingAddress from '../../containers/CopyMailingAddress';
+import CopyMailingAddress from 'vet360/containers/CopyMailingAddress';
 import AddressForm from './AddressForm';
 
 export default class AddressEditModal extends React.Component {
